✨: move buy-required action into Process component
what: relocate Buy required items from process page wrapper.

why: make the action available wherever Process is embedded.

how to test: run vitest specs for Process and Processes routes.
