+++ frontend/src/components/svelte/Process.svelte
@@ -18,9 +18,11 @@
     import { syncGameStateFromLocalIfStale } from '../../utils/gameState/common.js';
     import processes from '../../generated/processes.json';
     import { durationInSeconds } from '../../utils.js';
+    import items from '../../pages/inventory/json/items';
+    import { getPriceStringComponents } from '../../utils.js';
     import Chip from './Chip.svelte';
     import CompactItemList from './CompactItemList.svelte';
-    import { getItemCounts } from '../../utils/gameState/inventory.js';
+    import { buyItems, getItemCount, getItemCounts } from '../../utils/gameState/inventory.js';
     import { getItemMetadata } from './compactItemListHelpers.js';
     import { getItemMap } from '../../utils/itemResolver.js';
     import { initializeQaCheats, qaCheatsAvailability, qaCheatsEnabled } from '../../lib/qaCheats';
@@ -59,10 +61,187 @@
     let requirementItemMap = new Map();
     let requirementItemRequestId = 0;
     let previousRequirementKey = '';
+    let disableBuyRequired = true;
+    let disabledBuyRequiredReason = 'No required items are purchasable.';
+    let buyRequiredToastVisible = false;
+    let buyRequiredToastMessage = '';
+    const disabledBuyRequiredReasonId = `buy-required-disabled-reason-${processId}`;
+    const QUANTITY_PRECISION = 1_000_000;
+    const CURRENCY_EPSILON = 1e-9;
+    const defaultCurrencyItem = items.find((item) => item.name === 'dUSD') ?? null;
 
     // Slightly longer than the 1s CSS animation to avoid timing races.
     const pulseDurationMs = 1050;
     const updateIntervalMs = 100;
+    const roundDownQuantity = (value) => {
+        if (!Number.isFinite(value) || value <= 0) {
+            return 0;
+        }
+        return Math.floor((value + CURRENCY_EPSILON) * QUANTITY_PRECISION) / QUANTITY_PRECISION;
+    };
+    const roundCurrency = (value) => {
+        if (!Number.isFinite(value)) {
+            return 0;
+        }
+        return Math.round(value * QUANTITY_PRECISION) / QUANTITY_PRECISION;
+    };
+
+    const getCurrencyItem = (symbol) => {
+        if (!symbol) {
+            return defaultCurrencyItem;
+        }
+        return items.find((item) => item.name === symbol) ?? defaultCurrencyItem;
+    };
+
+    const getUnitPrice = (item) => {
+        const { price, symbol } = getPriceStringComponents(item?.price);
+        if (!Number.isFinite(price) || price <= 0) {
+            return null;
+        }
+
+        const currencyItem = getCurrencyItem(symbol);
+        if (!currencyItem) {
+            return null;
+        }
+
+        return {
+            unitPrice: price,
+            currencySymbol: symbol || defaultCurrencyItem?.name || 'dUSD',
+            currencyId: currencyItem.id,
+        };
+    };
+
+    const getPendingBuyRequirements = () => {
+        if (!process?.requireItems?.length) {
+            return [];
+        }
+
+        return process.requireItems
+            .map((req) => {
+                const have = getItemCount(req.id);
+                const neededQuantity = roundDownQuantity(req.count - have);
+                if (neededQuantity <= 0) {
+                    return null;
+                }
+
+                const item = items.find((i) => i.id === req.id);
+                const pricing = getUnitPrice(item);
+                if (!pricing) {
+                    return null;
+                }
+
+                return {
+                    id: req.id,
+                    quantity: neededQuantity,
+                    unitPrice: pricing.unitPrice,
+                    currencySymbol: pricing.currencySymbol,
+                    currencyId: pricing.currencyId,
+                };
+            })
+            .filter(Boolean);
+    };
+
+    const buildPurchasePlan = (pendingBuys) => {
+        const sortedRequirements = [...pendingBuys].sort(
+            (a, b) => a.unitPrice * a.quantity - b.unitPrice * b.quantity
+        );
+        const remainingByCurrency = sortedRequirements.reduce((acc, requirement) => {
+            if (acc[requirement.currencyId] == null) {
+                acc[requirement.currencyId] = roundCurrency(getItemCount(requirement.currencyId));
+            }
+            return acc;
+        }, {});
+
+        const purchasePlan = [];
+        let added = 0;
+        sortedRequirements.forEach((requirement) => {
+            const balance = remainingByCurrency[requirement.currencyId] ?? 0;
+            const affordableQuantity = roundDownQuantity(
+                (balance + CURRENCY_EPSILON) / requirement.unitPrice
+            );
+            if (affordableQuantity <= 0) {
+                return;
+            }
+
+            const quantity = roundDownQuantity(Math.min(requirement.quantity, affordableQuantity));
+            if (quantity <= 0) {
+                return;
+            }
+
+            const totalCost = roundCurrency(quantity * requirement.unitPrice);
+            if (totalCost > balance + CURRENCY_EPSILON) {
+                return;
+            }
+
+            purchasePlan.push({
+                id: requirement.id,
+                quantity,
+                price: requirement.unitPrice,
+                currencyId: requirement.currencyId,
+            });
+            remainingByCurrency[requirement.currencyId] = roundCurrency(balance - totalCost);
+            added = roundDownQuantity(added + quantity);
+        });
+
+        return {
+            purchasePlan,
+            added,
+        };
+    };
+
+    const getDisabledReason = () => {
+        if (!process || !process.requireItems) {
+            return 'No required items are purchasable.';
+        }
+
+        const missingRequirements = process.requireItems.filter(
+            (req) => roundDownQuantity(req.count - getItemCount(req.id)) > 0
+        );
+        if (missingRequirements.length === 0) {
+            return 'All required items are already available.';
+        }
+
+        const pendingBuys = getPendingBuyRequirements();
+        if (pendingBuys.length === 0) {
+            return 'Required items cannot be purchased.';
+        }
+
+        const { purchasePlan } = buildPurchasePlan(pendingBuys);
+        if (purchasePlan.length === 0) {
+            return 'Not enough currency to buy any still-needed required items.';
+        }
+
+        return '';
+    };
+
+    const updateBuyRequiredDisabledState = () => {
+        disabledBuyRequiredReason = getDisabledReason();
+        disableBuyRequired = Boolean(disabledBuyRequiredReason);
+    };
+
+    const buyRequiredItems = () => {
+        if (!process) return;
+
+        const pendingBuys = getPendingBuyRequirements();
+        if (pendingBuys.length === 0) {
+            updateBuyRequiredDisabledState();
+            return;
+        }
+
+        const { purchasePlan, added } = buildPurchasePlan(pendingBuys);
+        if (purchasePlan.length === 0) {
+            updateBuyRequiredDisabledState();
+            return;
+        }
+
+        buyItems(purchasePlan);
+        if (added > 0) {
+            buyRequiredToastMessage = `✓ Added ${added} items to inventory`;
+            buyRequiredToastVisible = true;
+            setTimeout(() => (buyRequiredToastVisible = false), 3000);
+        }
+        updateBuyRequiredDisabledState();
+    };
 
     const releaseItemImages = (items) => {
         items.forEach((item) => item?.releaseImage?.());
@@ -454,6 +633,10 @@
             void loadRequirementItemMap(requirementItems);
         }
     }
+
+    $: if (mounted) {
+        updateBuyRequiredDisabledState();
+    }
 </script>
 
 {#if mounted && process}
@@ -505,6 +688,27 @@
             <h4>Duration: {process.duration}</h4>
 
             {#if state === ProcessStates.NOT_STARTED}
+                <span
+                    class="buy-required-wrapper"
+                    title={disableBuyRequired ? disabledBuyRequiredReason : undefined}
+                >
+                    <button
+                        class="primary"
+                        type="button"
+                        on:click={buyRequiredItems}
+                        disabled={disableBuyRequired}
+                        aria-describedby={disableBuyRequired
+                            ? disabledBuyRequiredReasonId
+                            : undefined}
+                    >
+                        Buy required items
+                    </button>
+                    {#if disableBuyRequired}
+                        <span class="sr-only" id={disabledBuyRequiredReasonId}>
+                            {disabledBuyRequiredReason}
+                        </span>
+                    {/if}
+                </span>
                 <div
                     class="start-action"
                     class:pulse={isPulsing}
@@ -583,6 +787,9 @@
             {/if}
         </div>
     </Chip>
+    {#if buyRequiredToastVisible}
+        <div class="toast" role="status" aria-live="polite">{buyRequiredToastMessage}</div>
+    {/if}
 {:else if mounted}
     <div class="process-error">Unknown process.</div>
 {/if}
@@ -645,6 +852,55 @@
         margin: 0;
         font-size: 0.9rem;
         color: #fff7ed;
+        text-align: center;
+    }
+
+    .buy-required-wrapper {
+        position: relative;
+        display: inline-block;
+        align-self: center;
+    }
+
+    .primary {
+        background-color: #2f5b2f;
+        color: white;
+        border: none;
+        border-radius: 6px;
+        padding: 8px 16px;
+        cursor: pointer;
+    }
+
+    .primary:focus-visible {
+        outline: 2px solid #fff;
+        outline-offset: 2px;
+    }
+
+    .primary[disabled] {
+        opacity: 0.6;
+        cursor: not-allowed;
+    }
+
+    .sr-only {
+        position: absolute;
+        width: 1px;
+        height: 1px;
+        padding: 0;
+        margin: -1px;
+        overflow: hidden;
+        clip: rect(0, 0, 0, 0);
+        white-space: nowrap;
+        border: 0;
+    }
+
+    .toast {
+        position: fixed;
+        bottom: 20px;
+        left: 50%;
+        transform: translateX(-50%);
+        background-color: #cacaca;
+        color: #fff;
+        padding: 10px 20px;
+        border-radius: 5px;
         text-align: center;
     }
